+++ src/entities/data-bubbles/lib/drawer/DrawableDataBubble.ts
