+++ src/entities/data-bubbles/lib/drawer/drawBuble.ts
