+++ src/entities/data-bubbles/lib/drawer/getFunctionGetColorByDelta.ts
