+++ src/entities/data-bubbles/lib/index.ts
@@ -1,5 +1,1 @@
-export * from "./getFunctionGetColorByDelta";
-
-export * from "./DrawableDataBubble";
-
-export * from "./DrawerDataBubbles";
+export * from "./drawer/DrawerDataBubbles";
